pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.119.0",
    "mysqlclient>=2.2.7",
    "pydantic-settings>=2.0.0",
    "sqlmodel>=0.0.27",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
    "uvicorn>=0.38.0",
]
